Updated the module header of iss.py
Initially, the module header was import urllib2. However, urllib2 is only available for Python 2. In Python 3, urllib2 was separated into two different modules: urllib.request and urllib.error, meaning that I will have to erase the urllib2 module and replace it with urllib.request along with the methods that follows it.